+++ project_stats/project_stats.py
@@ -15,6 +15,10 @@
 #
 # Source url (https://github.com/LuqueDaniel/ninja-project-stats)
 
+"""
+    This module contain the main function of the plugin.
+"""
+
 #NINJA-IDE imports
 from ninja_ide.core import plugin
 
@@ -30,13 +34,21 @@ from PyQt4.QtGui import QTableWidget
 from PyQt4.QtGui import QAbstractItemView
 
 #PROJECT-STATS imports
-from get_stats import getStats
+from .get_stats import getStats
 
 
-class projectStatDialog(QDialog):
+class projectStatsDialog(QDialog):
+    """This class show project stats in a QDialog.
+
+    init Parameters:
+        projectInfo: Information of the current project.
+
+    Attributes:
+        projectStats: Contain stats of the project.
+    """
 
     def __init__(self, projectInfo):
-        super(projectStatDialog, self).__init__()
+        super(projectStatsDialog, self).__init__()
         self.setWindowTitle('Project Stats - %s' % projectInfo.name)
         self.setMinimumSize(500, 400)
         self.setMaximumSize(0, 0)
@@ -45,67 +57,74 @@ class projectStatDialog(QDialog):
         self.projectStats = getStats(projectInfo.path)
 
         #tabMenu
-        self.tabMenu = QTabWidget()
+        tabMenu = QTabWidget()
 
         #LAYOUTS
         #layoutTab
-        self.layoutTab = QVBoxLayout()
-        self.layoutTab.addWidget(self.tabMenu)
+        layoutTab = QVBoxLayout()
+        layoutTab.addWidget(tabMenu)
 
         #==Add layoutTabGeneral
-        self.layoutTabGeneral = QVBoxLayout()
-        self.layoutTabGeneral.addWidget(QLabel('Number of folders: %i' %
+        layoutTabGeneral = QVBoxLayout()
+        layoutTabGeneral.addWidget(QLabel('Number of folders: %i' %
                                     self.projectStats.info['numberFolders']))
-        self.layoutTabGeneral.addWidget(QLabel('Number of files: %i' %
+        layoutTabGeneral.addWidget(QLabel('Number of files: %i' %
                                     self.projectStats.info['numberFiles']))
-        self.layoutTabGeneral.addWidget(QLabel('Total number of lines: %i' %
+        layoutTabGeneral.addWidget(QLabel('Total number of lines: %i' %
                                     self.projectStats.info['numberLines']))
 
         #Add table fileTabGeneral at layoutTabGeneral
-        self.fileTableGeneral = QTableWidget(0, 2)
-        self._configTable(self.fileTableGeneral, 'generalFilesLines')
-        self.layoutTabGeneral.addWidget(self.fileTableGeneral)
+        fileTableGeneral = QTableWidget(0, 2)
+        self._configTable(fileTableGeneral, 'generalFilesLines')
+        layoutTabGeneral.addWidget(fileTableGeneral)
 
         #add widget tabGeneral at tabMenu
-        self.tabGeneral = QWidget()
-        self.tabGeneral.setLayout(self.layoutTabGeneral)
-        self.tabMenu.addTab(self.tabGeneral, 'General')
+        tabGeneral = QWidget()
+        tabGeneral.setLayout(layoutTabGeneral)
+        tabMenu.addTab(tabGeneral, 'General')
 
         #==Add layoutTabPy
         #if project contain py files add a py tab
         if self.projectStats.info['numberPyFiles'] != 0:
-            self.layoutTabPy = QVBoxLayout()
-            self.layoutTabPy.addWidget(QLabel('Number of .py files: %i' %
+            layoutTabPy = QVBoxLayout()
+            layoutTabPy.addWidget(QLabel('Number of .py files: %i' %
                                     self.projectStats.info['numberPyFiles']))
-            self.layoutTabPy.addWidget(QLabel('Number of .pyc files: %i' %
+            layoutTabPy.addWidget(QLabel('Number of .pyc files: %i' %
                                     self.projectStats.info['numberPycFiles']))
-            self.layoutTabPy.addWidget(QLabel('Total number of lines: %i' %
+            layoutTabPy.addWidget(QLabel('Total number of lines: %i' %
                                     self.projectStats.info['numberPyLines']))
 
             #add table fileTablelist at layoutTabPy
-            self.fileTablePy = QTableWidget(10, 2)
-            self._configTable(self.fileTablePy, 'pyFilesLines')
-            self.layoutTabPy.addWidget(self.fileTablePy)
+            fileTablePy = QTableWidget(10, 2)
+            self._configTable(fileTablePy, 'pyFilesLines')
+            layoutTabPy.addWidget(fileTablePy)
 
             #add Widget TabPy at tabMenu
-            self.tabPy = QWidget()
-            self.tabPy.setLayout(self.layoutTabPy)
-            self.tabMenu.addTab(self.tabPy, '.py')
+            tabPy = QWidget()
+            tabPy.setLayout(layoutTabPy)
+            tabMenu.addTab(tabPy, '.py')
 
         #Vertical Layout
-        self.vLayout = QVBoxLayout(self)
-        self.vLayout.setContentsMargins(15, 10, 15, 10)
+        vLayout = QVBoxLayout(self)
+        vLayout.setContentsMargins(15, 10, 15, 10)
         #add label with project name
-        self.vLayout.addWidget(QLabel('<b>Project name:</b> %s' %
+        vLayout.addWidget(QLabel('<b>Project name:</b> %s' %
                                 projectInfo.name))
         #add tabMenu
-        self.vLayout.addLayout(self.layoutTab)
+        vLayout.addLayout(layoutTab)
 
     def _configTable(self, table, dictKey):
-        self.tableHeaders = ('Name', 'Number of lines')
+        """This function configure a table.
+
+        Parameters:
+            table: Table to configure.
+            dictKey: The dictKey.
+        """
+
+        self.tableHeaders = ('Path & File name', 'Number of lines')
         table.setRowCount(len(self.projectStats.info[dictKey]))
         table.setHorizontalHeaderLabels(self.tableHeaders)
-        #No edit items
+        #Disable edit items
         table.setEditTriggers(QAbstractItemView.NoEditTriggers)
         #Single selection
         table.setSelectionMode(QTableWidget.SingleSelection)
@@ -118,27 +137,36 @@ class projectStatDialog(QDialog):
 
         row = 0
         for item in list(self.projectStats.info[dictKey].items()):
-            table.setItem(row, 0, QTableWidgetItem(item[1]['name']))
+            table.setItem(row, 0, QTableWidgetItem(item[1]['pathInProject']))
             table.setItem(row, 1, QTableWidgetItem(str(item[1]['lines'])))
             row += 1
 
 
 class projectStatsMain(plugin.Plugin):
+    """Main class of the plugin.
+
+    Attributes:
+        ex_locator: ninja-ide explorer service.
+    """
 
     def initialize(self):
+        """This function start plugin"""
+
         #Create plugin menu
-        self.menu = QMenu()
-        self.menu.setTitle('Project Stats')
-        self.menu.addAction('Project Stats', lambda: self.projectStatAction())
+        menu = QMenu()
+        menu.setTitle('Project Stats')
+        menu.addAction('Project Stats', lambda: self.projectStatAction())
 
         #Add Project Stats menu
         self.ex_locator = self.locator.get_service('explorer')
-        self.ex_locator.add_project_menu(self.menu)
+        self.ex_locator.add_project_menu(menu)
 
     def projectStatAction(self):
+        """Init projectStatsDialog"""
+
         #Get project properties
         self.currentProject = self.ex_locator.get_tree_projects()._get_project_root()
 
         #Instance projectStatDialog
-        self.projectStatDialog = projectStatDialog(self.currentProject)
-        self.projectStatDialog.show()
+        self.projectStatsDialog = projectStatsDialog(self.currentProject)
+        self.projectStatsDialog.show()